Add boring-registry image.
I wrote a helm test for this but dropped it after. The upstream image has a helm chart
but fails upon deployment without GCP or S3 credentials as well. The image is dead simple:

https://github.com/TierMobility/boring-registry/blob/main/Dockerfile

so I think the risk of this failing with GCS/S3 is low enough that it's not worth the complexity
of adding a 3p service to our tests.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>