fix: check for correct names in ExternalDNS tests
Check for the correct deployment and service names in the ExternalDNS tests. It
appears that the Helm chart has changed to not append coredns- as a prefix anymore.
As a result, because the shell script sets both errexit and pipefail, the whole
test fails because the cluster objects are not found.

Signed-off-by: Mauren Berti <mauren.berti@chainguard.dev>